+++ libheif/codecs/avc_boxes.cc
@@ -351,7 +351,7 @@ Error parse_sps_for_avcC_configuration(const uint8_t* sps, size_t size,
   size = sps_no_emul.size();
 
 
-  BitReader reader(sps, (int) size);
+  BitReader reader(sps, size);
 
   // skip NAL header
   reader.skip_bits(8);
